requirements.txt
mido==1.3.2
numpy==2.1.2
PyAudio==0.2.14
pyserial==3.5
